Tracks interesting goroutines by ID instead of stacktrace (#14)
* Tracks interesting goroutines by ID instead of stacktrace

Previously the leak detection compared the goroutine stacks from
the starting of a function with their stacks at the end to find
new goroutines. But, when there are preexisting goroutines whose
stacks change during the function, they are incorrectly detected
as a leaked goroutine. This update compare goroutine IDs instead.

* Changes interestingGoroutine to return an error