refactor: extract 'LeakCheckConfiguration'
This struct stores the slice of 'RoutinesSafeToIgnore' that are used as filters
in `interestingGoroutines`.

All package public functions for checking now immediately delegate to the
`DefaultCheckConfiguration` which has the previous hard-coded list of values.

This provides an extension point where a consumer of the library can extend the
routines which are safe to ignore (i.e. explicitly allow a certain goroutine to
leak) by either copying the DefaultCheckConfiguration and adding their own
values (ideal), or by simply mutating the default value (less ideal).